Update Application.php
Add an additional event 'beforeLoadModules' to handleRequest. 
My personal use case was to hide table_tabs in the top left menu as described here: http://xataface.com/forum/viewtopic.php?t=4445
But with the g2 module this could not work, as the beforeHandleRequest function gets called after the loading of the module. Thus the new event.